Revert "[Harness] Refactor simulators and devices listing. (#8064)"
This reverts commit e53b686ef63705f2381a83adfb3a5bcfac88eb73.
