+++ tests/xharness/AppRunner.cs
@@ -8,8 +8,6 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Execution;
 using Xharness.Jenkins.TestTasks;
 using Xharness.Listeners;
@@ -73,8 +71,8 @@ namespace Xharness {
 
 		public ILog MainLog { get; set; }
 
-		public ISimulatorDevice [] Simulators { get; set; }
-		ISimulatorDevice simulator => Simulators [0];
+		public SimDevice [] Simulators { get; set; }
+		SimDevice simulator => Simulators [0];
 
 		public string BundleIdentifier { get; private set; }
 
@@ -150,7 +148,7 @@ namespace Xharness {
 			}
 
 			var selected = devs.ConnectedDevices.Where ((v) => deviceClasses.Contains (v.DeviceClass) && v.IsUsableForDebugging != false);
-			IHardwareDevice selected_data;
+			Device selected_data;
 			if (selected.Count () == 0) {
 				throw new Exception ($"Could not find any applicable devices with device class(es): {string.Join (", ", deviceClasses)}");
 			} else if (selected.Count () > 1) {
