+++ tests/xharness/Jenkins/TestTasks/AggregatedRunSimulatorTask.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using Xharness.Hardware;
 
 namespace Xharness.Jenkins.TestTasks
 {
