+++ tests/xharness/Jenkins/TestTasks/RunDeviceTask.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Xharness.Hardware;
 
 namespace Xharness.Jenkins.TestTasks
 {
-	class RunDeviceTask : RunXITask<IHardwareDevice>
+	class RunDeviceTask : RunXITask<Device>
 	{
 		AppInstallMonitorLog install_log;
 		public override string ProgressMessage {
@@ -33,7 +32,7 @@ namespace Xharness.Jenkins.TestTasks
 			}
 		}
 
-		public RunDeviceTask (MSBuildTask build_task, IEnumerable<IHardwareDevice> candidates)
+		public RunDeviceTask (MSBuildTask build_task, IEnumerable<Device> candidates)
 			: base (build_task, candidates.OrderBy ((v) => v.DebugSpeed))
 		{
 			switch (build_task.Platform) {
