+++ tests/xharness/MonoNativeInfo.cs
@@ -26,9 +26,6 @@
 using System;
 using System.IO;
 using System.Xml;
-
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Utilities;
 
 namespace Xharness
