+++ tests/xharness/Targets/TVOSTarget.cs
@@ -1,8 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Xml;
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Utilities;
 
 namespace Xharness.Targets
