+++ tests/xharness/Targets/WatchOSTarget.cs
@@ -4,9 +4,6 @@ using System.IO;
 using System.Xml;
 using Xharness.Utilities;
 
-using Xamarin;
-using Xharness.Hardware;
-
 namespace Xharness.Targets
 {
 	public class WatchOSTarget : iOSTarget
